# generator/api-docs.json
{"swagger":"2.0","info":{"description":"Access to PokeBible referential with REST API","version":"1.0.0","title":"PokeBible API","contact":{"email":"fizou@yahoo.com"},"license":{"name":"Apache 2.0","url":"http://www.apache.org/licenses/LICENSE-2.0.html"}},"host":"localhost:8085","basePath":"/","tags":[{"name":"Pokemon Rest Api","description":"Rest Controllers"}],"paths":{"/api/generateToken":{"post":{"tags":["Pokemon Rest Api"],"summary":"generateToken","operationId":"generateTokenUsingPOST","consumes":["application/json"],"produces":["application/json"],"parameters":[{"in":"body","name":"credentials","description":"credentials","required":true,"schema":{"$ref":"#/definitions/ApiCredentials"}}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/ApiResponse"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"security":[{"JWT Token":["global"]}]}},"/api/pokemons":{"get":{"tags":["Pokemon Rest Api"],"summary":"findAll","operationId":"findAllUsingGET","produces":["*/*"],"parameters":[{"name":"page","in":"query","description":"page","required":false,"type":"integer","format":"int32"},{"name":"searchString","in":"query","description":"searchString","required":false,"type":"string"},{"name":"size","in":"query","description":"size","required":false,"type":"integer","format":"int32"},{"name":"sortDirection","in":"query","description":"sortDirection","required":false,"type":"string"},{"name":"sortField","in":"query","description":"sortField","required":false,"type":"string"}],"responses":{"200":{"description":"OK","schema":{"type":"array","items":{"$ref":"#/definitions/PokemonRes"}}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"security":[{"JWT Token":["global"]}]},"post":{"tags":["Pokemon Rest Api"],"summary":"save","operationId":"saveUsingPOST","consumes":["application/json"],"produces":["*/*"],"parameters":[{"in":"body","name":"pokemon","description":"pokemon","required":true,"schema":{"$ref":"#/definitions/PokemonReq"}}],"responses":{"201":{"description":"Created","schema":{"$ref":"#/definitions/PokemonRes"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"security":[{"JWT Token":["global"]}]}},"/api/pokemons/{id}":{"get":{"tags":["Pokemon Rest Api"],"summary":"findById","operationId":"findByIdUsingGET","produces":["*/*"],"parameters":[{"name":"id","in":"path","description":"id","required":true,"type":"integer","format":"int64"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/PokemonRes"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"security":[{"JWT Token":["global"]}]},"put":{"tags":["Pokemon Rest Api"],"summary":"saveOrUpdate","operationId":"saveOrUpdateUsingPUT","consumes":["application/json"],"produces":["*/*"],"parameters":[{"name":"id","in":"path","description":"id","required":true,"type":"integer","format":"int64"},{"in":"body","name":"pokemon","description":"pokemon","required":true,"schema":{"$ref":"#/definitions/PokemonReq"}}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/PokemonRes"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"security":[{"JWT Token":["global"]}]},"delete":{"tags":["Pokemon Rest Api"],"summary":"delete","operationId":"deleteUsingDELETE","produces":["*/*"],"parameters":[{"name":"id","in":"path","description":"id","required":true,"type":"integer","format":"int64"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/ApiResponse"}},"204":{"description":"No Content"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"}},"security":[{"JWT Token":["global"]}]},"patch":{"tags":["Pokemon Rest Api"],"summary":"update","operationId":"updateUsingPATCH","consumes":["application/json"],"produces":["*/*"],"parameters":[{"name":"id","in":"path","description":"id","required":true,"type":"integer","format":"int64"},{"in":"body","name":"map","description":"map","required":true,"schema":{"type":"object","additionalProperties":{"type":"string"}}}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/PokemonRes"}},"204":{"description":"No Content"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"}},"security":[{"JWT Token":["global"]}]}},"/api/test":{"get":{"tags":["Pokemon Rest Api"],"summary":"testApi","operationId":"testApiUsingGET","produces":["*/*"],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/ApiResponse"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"security":[{"JWT Token":["global"]}]}}},"securityDefinitions":{"JWT Token":{"type":"apiKey","name":"Authorization","in":"header"}},"definitions":{"ApiCredentials":{"type":"object","properties":{"password":{"type":"string"},"username":{"type":"string"}},"title":"ApiCredentials"},"ApiResponse":{"type":"object","properties":{"debug":{"type":"string"},"message":{"type":"string"},"path":{"type":"string"},"status":{"type":"string"},"statusLabel":{"type":"string"},"timestamp":{"type":"string"},"version":{"type":"string"}},"title":"ApiResponse"},"PokemonReq":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"},"number":{"type":"string","pattern":"[0-9]+"},"type1":{"type":"string","pattern":"(?:(?:^|, )(NORMAL|GRASS|FIRE|WATER|FIGHTING|FLYING|POISON|GROUND|ROCK|BUG|GHOST|ELECTRIC|PSYCHIC|ICE|DRAGON|DARK|STEEL|FAIRY))+$"},"type2":{"type":"string","pattern":"(?:(?:^|, )(NONE|NORMAL|GRASS|FIRE|WATER|FIGHTING|FLYING|POISON|GROUND|ROCK|BUG|GHOST|ELECTRIC|PSYCHIC|ICE|DRAGON|DARK|STEEL|FAIRY))+$"}},"title":"PokemonReq"},"PokemonRes":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"},"number":{"type":"string","pattern":"[0-9]+"},"pictureUrl":{"type":"string"},"type":{"type":"string"},"type1":{"type":"string","pattern":"(?:(?:^|, )(NORMAL|GRASS|FIRE|WATER|FIGHTING|FLYING|POISON|GROUND|ROCK|BUG|GHOST|ELECTRIC|PSYCHIC|ICE|DRAGON|DARK|STEEL|FAIRY))+$"},"type1Label":{"type":"string"},"type1PictureUrl":{"type":"string"},"type2":{"type":"string","pattern":"(?:(?:^|, )(NONE|NORMAL|GRASS|FIRE|WATER|FIGHTING|FLYING|POISON|GROUND|ROCK|BUG|GHOST|ELECTRIC|PSYCHIC|ICE|DRAGON|DARK|STEEL|FAIRY))+$"},"type2Label":{"type":"string"},"type2PictureUrl":{"type":"string"}},"title":"PokemonRes"}}}